+++ src/main/java/dinkplugin/notifiers/SpeedrunNotifier.java
@@ -9,6 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.events.WidgetLoaded;
 import net.runelite.api.widgets.Widget;
 import org.apache.commons.lang3.StringUtils;
+import org.jetbrains.annotations.VisibleForTesting;
 
 import java.time.Duration;
 import java.util.regex.Matcher;
@@ -17,10 +18,10 @@ import java.util.regex.Pattern;
 @Slf4j
 public class SpeedrunNotifier extends BaseNotifier {
     private static final Pattern TIME_PATTERN = Pattern.compile("(?<minutes>\\d+):(?<seconds>\\d{2})\\.(?<fractional>\\d{2})");
-    private static final int SPEEDRUN_COMPLETED_GROUP_ID = 781;
-    private static final int SPEEDRUN_COMPLETED_QUEST_NAME_CHILD_ID = 4;
-    private static final int SPEEDRUN_COMPLETED_DURATION_CHILD_ID = 10;
-    private static final int SPEEDRUN_COMPLETED_PB_CHILD_ID = 12;
+    static final @VisibleForTesting int SPEEDRUN_COMPLETED_GROUP_ID = 781;
+    static final @VisibleForTesting int SPEEDRUN_COMPLETED_QUEST_NAME_CHILD_ID = 4;
+    static final @VisibleForTesting int SPEEDRUN_COMPLETED_DURATION_CHILD_ID = 10;
+    static final @VisibleForTesting int SPEEDRUN_COMPLETED_PB_CHILD_ID = 12;
 
     @Override
     public boolean isEnabled() {