+++ src/test/java/dinkplugin/notifiers/KillCountNotifierTest.java
@@ -1,0 +1,90 @@
+package dinkplugin.notifiers;
+
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.notifiers.data.BossNotificationData;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+class KillCountNotifierTest extends MockedNotifierTest {
+
+    @InjectMocks
+    KillCountNotifier notifier;
+
+    @Override
+    @BeforeEach
+    protected void setUp() {
+        super.setUp();
+
+        // init config mocks
+        when(config.notifyKillCount()).thenReturn(true);
+        when(config.killCountSendImage()).thenReturn(true);
+        when(config.killCountMessage()).thenReturn("%USERNAME% has defeated %BOSS% with a completion count of %COUNT%");
+    }
+
+    @Test
+    void testNotifyInterval() {
+        // more config
+        when(config.killCountNotifyInitial()).thenReturn(false);
+        when(config.killCountInterval()).thenReturn(70);
+
+        // fire event
+        String gameMessage = "Your King Black Dragon kill count is: 420.";
+        notifier.onGameMessage(gameMessage);
+
+        // check notification
+        verify(messageHandler).createMessage(
+            true,
+            NotificationBody.builder()
+                .content(PLAYER_NAME + " has defeated King Black Dragon with a completion count of 420")
+                .extra(new BossNotificationData("King Black Dragon", 420, gameMessage))
+                .playerName(PLAYER_NAME)
+                .type(NotificationType.KILL_COUNT)
+                .build()
+        );
+    }
+
+    @Test
+    void testNotifyInitial() {
+        // more config
+        when(config.killCountNotifyInitial()).thenReturn(true);
+        when(config.killCountInterval()).thenReturn(99);
+
+        // fire event
+        String gameMessage = "Your King Black Dragon kill count is: 1.";
+        notifier.onGameMessage(gameMessage);
+
+        // check notification
+        verify(messageHandler).createMessage(
+            true,
+            NotificationBody.builder()
+                .content(PLAYER_NAME + " has defeated King Black Dragon with a completion count of 1")
+                .extra(new BossNotificationData("King Black Dragon", 1, gameMessage))
+                .playerName(PLAYER_NAME)
+                .type(NotificationType.KILL_COUNT)
+                .build()
+        );
+    }
+
+    @Test
+    void testIgnore() {
+        // more config
+        when(config.killCountNotifyInitial()).thenReturn(false);
+        when(config.killCountInterval()).thenReturn(99);
+
+        // fire event
+        String gameMessage = "Your King Black Dragon kill count is: 1.";
+        notifier.onGameMessage(gameMessage);
+
+        // ensure no message
+        verify(messageHandler, never()).createMessage(anyBoolean(), any());
+    }
+
+}